Added category edit option using axios and removed category delete option

+++ public/admin/assets/js/productManage.js
@@ -209,23 +209,25 @@ document.addEventListener("click", function (event) {
 
 //DELETE PRODUCT IMAGE
 async function deleteImage(productId, imageUrl) {
-  try {
-    const response = await axios.delete(
-      `/api/admin/product/delete-image?productId=${productId}&imageUrl=${imageUrl}`
-    );
-    if (response.data.success) {
-      console.log("product image deleted successfully");
-      // Remove image from table
-      const images = document.querySelectorAll("img");
-      images.forEach((img) => {
-        if (img.src.includes(imageUrl)) {
-          img.remove();
-        }
-      });
-    } else {
-      console.log("product image not deleted");
+  if (confirm("Are you sure you want to delete this image?")) {
+    try {
+      const response = await axios.delete(
+        `/api/admin/product/delete-image?productId=${productId}&imageUrl=${imageUrl}`
+      );
+      if (response.data.success) {
+        console.log("product image deleted successfully");
+        // Remove image from table
+        const images = document.querySelectorAll("img");
+        images.forEach((img) => {
+          if (img.src.includes(imageUrl)) {
+            img.remove();
+          }
+        });
+      } else {
+        console.log("product image not deleted");
+      }
+    } catch (err) {
+      console.log(err);
     }
-  } catch (err) {
-    console.log(err);
   }
 }
